move limit logic into own file
It is a bit complicated, so lets move it into its own file.
